requirements.txt
pymongo[srv]==4.2.0
from_root
pandas
numpy
scikit-learn
PyYAML
# -e .